test: add comprehensive test suite for database and utility save functions.
